pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ADMeth"
version = "0.1.0"
description = "Pipeline for methylation model evaluation"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "torch",
    "matplotlib",
    "seaborn",
    "pyyaml",
    "joblib",
    "statsmodels",
    "gseapy",
    "networkx"
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["cli", "steps_evaluate", "steps_preprocess", "steps_recscores", "steps_mlmodels", "steps_plots", "steps_stats"]

[project.scripts]
admeth = "cli:main"
